refactor(doors): function organization
